organized the files inside src/dependency-analysis
